pyproject.toml
[project]
name = "daily-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pgvector",
  "sqlalchemy>=2.0,<3.0",
  "psycopg>=3.1",
  "psycopg-binary>=3.1",
  "alembic>=1.13,<2.0",
  "pydantic>=2.0,<3.0",
  "pydantic-settings>=2.0,<3.0",
  "ollama>=0.6.1",
  "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = ["ipython", "pytest>=8.0", "ruff>=0.5"]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
